# standard-be/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Standard BE API
  version: 1.0.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ImageRef:
      type: object
      required: [ url ]
      properties:
        url: { type: string }
        alt: { type: string }
    EmbedPlatform:
      type: string
      enum: [ youtube, vimeo ]
    SpotifyType:
      type: string
      enum: [ track, album, playlist, episode, show ]
...
    BlockType:
      type: string
      enum: [ paragraph, heading, image, quote, video, embed, gallery, link, spotify ]
    Block:
      type: object
      required: [ type ]
      properties:
        type:
          $ref: '#/components/schemas/BlockType'
      discriminator:
        propertyName: type
        mapping:
          paragraph: '#/components/schemas/ParagraphBlock'
          heading: '#/components/schemas/HeadingBlock'
          image: '#/components/schemas/ImageBlock'
          quote: '#/components/schemas/QuoteBlock'
          video: '#/components/schemas/VideoBlock'
          embed: '#/components/schemas/EmbedBlock'
          gallery: '#/components/schemas/GalleryBlock'
          link: '#/components/schemas/LinkBlock'
          spotify: '#/components/schemas/SpotifyBlock'
    ParagraphBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ text ]
          properties:
            text:
              type: string
            html:
              type: string
    HeadingBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ level, text ]
          properties:
            level:
              type: integer
              minimum: 1
              maximum: 6
            text:
              type: string
    ImageBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ image ]
          properties:
            image:
              $ref: '#/components/schemas/ImageRef'
    QuoteBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ text ]
          properties:
            text:
              type: string
            author:
              type: string
    VideoBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ url ]
          properties:
            url: { type: string }
            poster: { type: string }
    EmbedBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ platform, id ]
          properties:
            platform: { $ref: '#/components/schemas/EmbedPlatform' }
            id: { type: string }
    GalleryBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ urls ]
          properties:
            urls:
              type: array
              items: { type: string }
    LinkBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ url ]
          properties:
            url: { type: string }
            title: { type: string }
            description: { type: string }
    SpotifyBlock:
      allOf:
        - $ref: '#/components/schemas/Block'
        - type: object
          required: [ url, spotifyType, spotifyId ]
          properties:
            url: { type: string }
            spotifyType: { $ref: '#/components/schemas/SpotifyType' }
            spotifyId: { type: string }
...
    EventStatus:
      type: string
      enum: [ DRAFT, PUBLISHED, ARCHIVED ]
    EventSummary:
      type: object
      required: [ id, title, slug, status, tags ]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { $ref: '#/components/schemas/EventStatus' }
        thumbnail: { $ref: '#/components/schemas/ImageRef' }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        timezone: { type: string }
        location: { type: string }
        ticketsUrl: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: string, format: uuid }
    EventDetails:
      allOf:
        - $ref: '#/components/schemas/EventSummary'
        - type: object
          required: [ content ]
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/Block' }
    PageEventSummary:
      type: object
      required: [ items, page, size ]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/EventSummary' }
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer }
...
    CreateEventRequest:
      type: object
      required: [ title, slug, content ]
      properties:
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        tags:
          type: array
          items: { type: string }
        thumbnail: { $ref: '#/components/schemas/ImageRef' }
        content:
          type: array
          items: { $ref: '#/components/schemas/Block' }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        timezone: { type: string }
        location: { type: string }
        ticketsUrl: { type: string }
    UpdateEventRequest:
      allOf:
        - $ref: '#/components/schemas/CreateEventRequest'
    ImportEventsRequest:
      type: object
      required: [ items ]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CreateEventRequest' }
paths:
...
  /public/events:
    get:
      tags: [ PublicEvents ]
      operationId: publicEventsList
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
      responses:
        "200":
          description: Events page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageEventSummary' }
  /public/events/{slug}:
    get:
      tags: [ PublicEvents ]
      operationId: publicEventGetBySlug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventDetails' }
...
  /admin/events:
    get:
      tags: [ AdminEvents ]
      operationId: adminEventList
      security:
        - bearerAuth: [ ]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/EventStatus' }
      responses:
        "200":
          description: Admin event page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageEventSummary' }
    post:
      tags: [ AdminEvents ]
      operationId: adminEventCreate
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEventRequest' }
      responses:
        "201":
          description: Created draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventDetails' }
...
  /admin/events/{id}:
    get:
      tags: [ AdminEvents ]
      operationId: adminEventGet
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventDetails' }
    put:
      tags: [ AdminEvents ]
      operationId: adminEventUpdate
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEventRequest' }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventDetails' }
    delete:
      tags: [ AdminEvents ]
      operationId: adminEventDelete
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
...
  /admin/events/{id}/publish:
    post:
      tags: [ AdminEvents ]
      operationId: adminEventPublish
      security:
        - bearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventDetails' }
...
  /admin/events/import:
    post:
      tags: [ AdminEvents ]
      operationId: adminEventImport
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportEventsRequest' }
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EventSummary' }
...
  /admin/events/reset:
    post:
      tags: [ AdminEvents ]
      operationId: adminEventReset
      security:
        - bearerAuth: [ ]
      responses:
        "204":
          description: Reset complete